pyproject.toml
[build-system]
requires = ["setuptools==75.6.0", "wheel==0.45.1"]
build-backend = "setuptools.build_meta"

[project]
name = "protocolquant"
version = "0.1.0"
description = "Protocol-driven, QC-gated image quantification"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "ProtocolQuant Contributors"}]
dependencies = [
  "numpy==2.1.3",
  "scipy==1.14.1",
  "pandas==2.2.3",
  "scikit-image==0.24.0",
  "pydantic==2.10.4",
  "tifffile==2024.9.20",
  "ome-types==0.5.3",
  "openpyxl==3.1.5",
  "jinja2==3.1.4",
  "matplotlib==3.9.2",
  "PyYAML==6.0.2",
]

[project.optional-dependencies]
segmentation-cellpose = [
  "cellpose==3.0.11",
  "torch==2.5.1",
]
ui = [
  "napari==0.5.6",
  "PyQt6==6.7.1",
]
dev = [
  "pytest==8.3.4",
  "pytest-cov==6.0.0",
  "black==24.10.0",
  "ruff==0.8.2",
  "pre-commit==4.0.1",
  "mypy==1.13.0",
]

[project.scripts]
protocolquant = "protocolquant.cli:main"

[tool.setuptools]
packages = ["protocolquant", "protocolquant.segmentation", "protocolquant.gui"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
ignore_missing_imports = true
strict_optional = true
